requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
pillow==10.2.0
svgwrite==1.4.3
svglib==1.5.1
reportlab==4.0.9 